+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Fix highlighting of `+`-prefixed options in help screens. Closes {issue}`316`.
+
 ## {gh}`3.4.0 (2022-12-01) <compare/v3.3.4...v3.4.0>`
 
 - Streamline setup of Sphinx extensions.